@@ -8,9 +8,9 @@
     "@testing-library/user-event": "^7.2.1",
     "formik": "^2.1.5",
     "pure-react-carousel": "^1.27.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "3.4.3",
     "styled-components": "^5.1.1",
     "yup": "^0.29.3"
