@@ -1,6 +1,9 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import {createRoot} from "react-dom";
 import App from "./App";
 import "./index.css";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById('root')
+const root = createRoot(container)
+root.render(<App />)
+// ReactDOM.render(<App />, document.getElementById("root"));
